Add navigation bar functionality and responsiveness
Discord button has been removed to aid in phone screen responsiveness.
Discord button will be moved to the socials page and the contact page.

Reformatted HTML and CSS

Colours are most likely not final
Pages in the nav menu are most likely not final (might add, rename, or
remove some in the future)